Refactor: Move PatchedConics into core folder
