+++ src/core/PatchedConics/FlightEvent/SOIArrival.js
