+++ src/core/PatchedConics/FlightEvent/SOIDeparture.js
