+++ src/core/PatchedConics/PatchedConics.js
@@ -1,7 +1,7 @@
-import Module from "../../core/Module";
-import Events from "../../core/Events";
-import EphemerisObject from "../../core/EphemerisObject";
-import Body from "../../core/Body";
+import Module from "../Module";
+import Events from "../Events";
+import EphemerisObject from "../EphemerisObject";
+import Body from "../Body";
 import PropagatorPatchedConics from "./PropagatorPatchedConics";
 import VisualFlightEventSOIArrival from "./visual/FlightEvent/SOIArrival";
 import VisualFlightEventSOIDeparture from "./visual/FlightEvent/SOIDeparture";
