+++ src/core/PatchedConics/visual/FlightEvent/SOIArrival.js
@@ -1,5 +1,5 @@
 
-import VisualFlightEventAbstract from "../../../../core/visual/FlightEvent/Abstract";
+import VisualFlightEventAbstract from "../../../visual/FlightEvent/Abstract";
 
 export default class VisualFlightEventSOIArrival extends VisualFlightEventAbstract
 {
